Extract selection metadata to its own class
