+++ client/jetbrains/src/main/java/com/sourcegraph/find/FindPopupPanel.java
@@ -1,9 +1,6 @@
 package com.sourcegraph.find;
 
-import com.intellij.icons.AllIcons;
 import com.intellij.openapi.Disposable;
-import com.intellij.openapi.actionSystem.KeyboardShortcut;
-import com.intellij.openapi.keymap.KeymapUtil;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.Splitter;
 import com.intellij.ui.OnePixelSplitter;
@@ -18,12 +15,8 @@ import com.sourcegraph.browser.SourcegraphJBCefBrowser;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import javax.swing.*;
 import java.awt.*;
-import java.awt.event.InputEvent;
-import java.awt.event.KeyEvent;
 import java.util.Date;
-import java.util.Objects;
 
 /**
  * Inspired by <a href="https://sourcegraph.com/github.com/JetBrains/intellij-community/-/blob/platform/lang-impl/src/com/intellij/find/impl/FindPopupPanel.java">FindPopupPanel.java</a>
@@ -32,9 +25,7 @@ public class FindPopupPanel extends JBPanel<FindPopupPanel> implements Disposabl
     private final SourcegraphJBCefBrowser browser;
     private final PreviewPanel previewPanel;
     private final BrowserAndLoadingPanel browserAndLoadingPanel;
-    private JLabel selectionMetadataLabel;
-    private JLabel externalLinkLabel;
-    private JLabel openShortcutLabel;
+    private final SelectionMetadataPanel selectionMetadataPanel;
     private Date lastPreviewUpdate;
 
     public FindPopupPanel(@NotNull Project project) {
@@ -47,8 +38,12 @@ public class FindPopupPanel extends JBPanel<FindPopupPanel> implements Disposabl
         Splitter splitter = new OnePixelSplitter(true, 0.5f, 0.1f, 0.9f);
         add(splitter, BorderLayout.CENTER);
 
+        selectionMetadataPanel = new SelectionMetadataPanel();
         previewPanel = new PreviewPanel(project);
-        JPanel bottomPanel = createBottomPanel();
+
+        BorderLayoutPanel bottomPanel = new BorderLayoutPanel();
+        bottomPanel.add(selectionMetadataPanel, BorderLayout.NORTH);
+        bottomPanel.add(previewPanel, BorderLayout.CENTER);
 
         browserAndLoadingPanel = new BrowserAndLoadingPanel();
         JSToJavaBridgeRequestHandler requestHandler = new JSToJavaBridgeRequestHandler(project, this);
@@ -68,27 +63,6 @@ public class FindPopupPanel extends JBPanel<FindPopupPanel> implements Disposabl
         splitter.setSecondComponent(bottomPanel);
 
         lastPreviewUpdate = new Date();
-    }
-
-    @NotNull
-    private BorderLayoutPanel createBottomPanel() {
-        BorderLayoutPanel bottomPanel = new BorderLayoutPanel();
-        JPanel selectionMetadataPanel = new JPanel(new FlowLayout(FlowLayout.LEFT, 5, 10));
-
-        selectionMetadataLabel = new JLabel();
-        externalLinkLabel = new JLabel("", AllIcons.Ide.External_link_arrow, SwingConstants.LEFT);
-        externalLinkLabel.setVisible(false);
-        KeyboardShortcut altEnterShortcut = new KeyboardShortcut(KeyStroke.getKeyStroke(KeyEvent.VK_ENTER, InputEvent.ALT_DOWN_MASK), null);
-        String altEnterShortcutText = KeymapUtil.getShortcutText(altEnterShortcut);
-        openShortcutLabel = new JLabel(altEnterShortcutText);
-        openShortcutLabel.setVisible(false);
-
-        selectionMetadataPanel.add(selectionMetadataLabel);
-        selectionMetadataPanel.add(externalLinkLabel);
-        selectionMetadataPanel.add(openShortcutLabel);
-        bottomPanel.add(selectionMetadataPanel, BorderLayout.NORTH);
-        bottomPanel.add(previewPanel, BorderLayout.CENTER);
-        return bottomPanel;
     }
 
     @Nullable
@@ -120,52 +94,22 @@ public class FindPopupPanel extends JBPanel<FindPopupPanel> implements Disposabl
     }
 
     public void indicateLoading() {
+        selectionMetadataPanel.clearSelectionMetadataLabel();
         previewPanel.setLoading(true);
         previewPanel.clearContent();
-        clearSelectionMetadataLabel();
     }
 
     public void setPreviewContent(@NotNull PreviewContent previewContent) {
+        selectionMetadataPanel.setSelectionMetadataLabel(previewContent);
         previewPanel.setContent(previewContent);
     }
 
     public void clearPreviewContent() {
+        selectionMetadataPanel.clearSelectionMetadataLabel();
         previewPanel.setContent(null);
     }
 
     public void setBrowserVisible(boolean visible) {
         browserAndLoadingPanel.setBrowserVisible(visible);
-    }
-
-    public void clearSelectionMetadataLabel() {
-        selectionMetadataLabel.setText("");
-        externalLinkLabel.setVisible(false);
-        openShortcutLabel.setVisible(false);
-    }
-
-    public void setSelectionMetadataLabel(@NotNull PreviewContent previewContent) {
-        String metadataText = getMetadataText(previewContent);
-        selectionMetadataLabel.setText(metadataText);
-        externalLinkLabel.setVisible(!previewContent.opensInEditor());
-        openShortcutLabel.setToolTipText("Press " + openShortcutLabel.getText() + " to open the selected file" +
-            (previewContent.opensInEditor() ? " in the editor." : " in your browser."));
-        openShortcutLabel.setVisible(true);
-    }
-
-    @NotNull
-    private String getMetadataText(@NotNull PreviewContent previewContent) {
-        if (Objects.equals(previewContent.getResultType(), "file") || Objects.equals(previewContent.getResultType(), "path")) {
-            return previewContent.getRepoUrl() + ":" + previewContent.getPath();
-        } else if (Objects.equals(previewContent.getResultType(), "repo")) {
-            return previewContent.getRepoUrl();
-        } else if (Objects.equals(previewContent.getResultType(), "symbol")) {
-            return previewContent.getSymbolName() + " (" + previewContent.getSymbolContainerName() + ")";
-        } else if (Objects.equals(previewContent.getResultType(), "diff")) {
-            return previewContent.getCommitMessagePreview() != null ? previewContent.getCommitMessagePreview() : "";
-        } else if (Objects.equals(previewContent.getResultType(), "commit")) {
-            return previewContent.getCommitMessagePreview() != null ? previewContent.getCommitMessagePreview() : "";
-        } else {
-            return "";
-        }
     }
 }
